Move helper functions into separate files
