+++ functions/src/metadata.ts
@@ -2,8 +2,10 @@
  * sri sri guru gaurangau jayatah
  */
 
+import { File } from '@google-cloud/storage';
 import { ContentDetails } from './ContentDetails';
 import { abbreviateLanguages, parseLanguages } from './languages';
+import { StorageFileMetadata } from './StorageFileMetadata';
 
 export function composeFileName(
   id: string,
@@ -32,5 +34,27 @@ export function composeMediaMetadata(
     'BVNM Archive ID': id,
     title: contentDetails.title,
     date: contentDetails.date?.substr(0, 4),
+  };
+}
+
+export function composeStorageMetadata(
+  id: string,
+  sourceFile: File,
+  contentDetails: ContentDetails
+): Partial<StorageFileMetadata> {
+  return {
+    /*
+     * File name may need encoding according to RFC 5987 as we have non-ASCII characters.
+     * https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/encodeURIComponent#encoding_for_content-disposition_and_link_headers
+     * https://stackoverflow.com/a/32782542/3082178
+     */
+    contentDisposition: `attachment; filename="${composeFileName(
+      id,
+      contentDetails
+    )}"`,
+    metadata: {
+      source: `${sourceFile.bucket.name}/${sourceFile.name}#${sourceFile.generation}`,
+      sourceMd5Hash: sourceFile.metadata.md5Hash,
+    },
   };
 }
